Let fsl_sub expose time limits instead of selecting a queue.
Instead of translating the time limit reported by the various FSL tools
into a specific queue, they get simply passed to SGE (via 'h_rt') to let
the cluster engine decide which queue is appropriate.

This has two advantages: 1) It doesn't require editing fsl_sub for each
cluster it is running on (to adjust to queue names and their limits) and
2) it allows for much more flexible, ressource-oriented SGE setups,
without binding a job to a specific queue a priori.